Add Id::read_packed_ids
One of git's little optimizations for packfile indices is that the index file contains
a sub-stream of all ids contained in the packfile in the order the objects can be
found in the packfile. They can be directly treated as a list of ids, straight from
disk.

So. Before I continue, I have to state:

This is not necessary, strictly speaking, but I wanted to make sure I understood how
to do it. That is, after all, the purpose of this repository.

In order to do this, we pre-allocate a vector of IDs using an object count, grab a
mutable slice from it, then pun that mutable slice of Ids to a mutable slice of u8.
We then read directly from the input into the id list.

As for why this isn't necessary: anecdotally, it doesn't look like this really
speeds anything up, though it may save us a copy of the vector data as its read
in. Packfile indices are read once at startup, and while a git repository may have
many packfiles, eventually those packfiles & indices are combined (See git-gc [1]
for more.) Anecdotally based on observed performance of `git_rs_log` against the
node repo, I don't think loading the index is the slow part of the program (We start
streaming loglines pretty rapidly).

That said, this points at another problem: I'm missing benchmarks. Those would tell
me where to optimize to catch up with Git. Anyway.

[1]: https://git-scm.com/docs/git-gc